Refactor core system, update client logic, and add transaction history
- Performed a complete refactoring of the main system architecture for better maintainability.
- Overhauled the client management system to improve data flow and structure.
- Implemented transaction history tracking for bank accounts.

+++ src/main/java/com/github/azcx1/banksystem/BankController.java
@@ -76,16 +76,16 @@ public class BankController {
         bankService.openSavingsAccount(currentClient, currency);
     }
 
-    public void deposit(double amount) {
+    public void deposit(BigDecimal amount) {
         bankService.deposit(
                 currentBankAccount.getAccountNumber(),
-                BigDecimal.valueOf(amount)
+                amount
         );
     }
-    public void withdraw(double amount) {
+    public void withdraw(BigDecimal amount) {
         bankService.withdraw(
                 currentBankAccount.getAccountNumber(),
-                BigDecimal.valueOf(amount)
+                amount
         );
     }
     public void transfer(String to, double amount) {
